% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{estimate_pi}
\alias{estimate_pi}
\title{Montecarlo estimation of Pi and its own Plot function}
\usage{
estimate_pi(B = 5000, seed = 10)
}
\arguments{
\item{B}{A  \code{numeric} (integer) used to denote the number of simulations.}

\item{seed}{A \code{numeric} (integer) used to control the seed of the random number
generator used by this function.}
}
\value{
A  \code{list} containing the following attributes:
\describe{
     \item{estimated_pi}{Estimated value of Pi}
     \item{points}{Data frame with 3 variables used to plot the area of the estimated Pi}
}
}
\description{
Compute an approximation of Pi by Monte-Carlo integration using
uniform sampling and Plot its result
}
\examples{
estimate_pi()
plot(estimate_pi())
}
